+++ packages/remme-public-key-storage/src/models/PublicKeyInfo.ts
@@ -7,6 +7,7 @@ export interface IPublicKeyInfo {
     valid_to: number;
     entity_hash: string;
     entity_hash_signature: string;
+    type: string;
 }
 
 /**
@@ -21,6 +22,7 @@ export class PublicKeyInfo {
     public validTo: number;
     public entityHash: string;
     public entityHashSignature: string;
+    public type: string;
 
     constructor(data: IPublicKeyInfo) {
         this.ownerPublicKey = data.owner_public_key;
@@ -31,5 +33,6 @@ export class PublicKeyInfo {
         this.validTo = data.valid_to;
         this.entityHash = data.entity_hash;
         this.entityHashSignature = data.entity_hash_signature;
+        this.type = data.type;
     }
 }