+++ packages/remme-utils/src/functions/index.ts
@@ -1,7 +1,8 @@
-import * as bytes from "utf8-bytes";
-import { createHash } from "crypto";
+import { createHash, randomBytes } from "crypto";
 import * as b64 from "base64-js";
 import * as forge from "node-forge";
+import * as secp256k1 from "secp256k1";
+import {PATTERNS} from "../../dist";
 
 export const sha512 = (value: Buffer | string) => createHash("sha512").update(value).digest("hex");
 
@@ -14,7 +15,7 @@ export const hexToBytes = (str: string) => {
     return new Uint8Array(arr);
 };
 
-export const bytesToHex = (uint8arr: Uint8Array): string => {
+export const bytesToHex = (uint8arr: Buffer | Uint8Array): string => {
     let hexStr: string = "";
     uint8arr.forEach((item) => {
         let hex = (item & 0xff).toString(16);
@@ -26,10 +27,6 @@ export const bytesToHex = (uint8arr: Uint8Array): string => {
 
 export const base64ToArrayBuffer = (base64: string): Uint8Array => {
    return b64.toByteArray(base64);
-};
-
-export const utf8ToBytes = (str: string): any => {
-    return bytes(str);
 };
 
 export const toHex = (str: string): string => {
@@ -44,10 +41,7 @@ export const generateAddress = (familyName: string, data: string): string => {
     return `${sha512(familyName).slice(0, 6)}${sha512(data).slice(0, 64)}`;
 };
 
-export const toHexString = (byteArray) => Array.from(byteArray, (byte: any) =>
-    ("0" + (byte & 0xFF).toString(16)).slice(-2)).join("");
-
-export const toUTF8Array = (str) => {
+export const toUTF8Array = (str: string): Buffer | Uint8Array => {
     const utf8 = [];
     for (let i = 0; i < str.length; i++) {
         let charCode = str.charCodeAt(i);
@@ -117,6 +111,22 @@ export const publicKeyFromPem = (publicKey: forge.pki.PEM): forge.pki.Certificat
     }
 };
 
+export const privateKeyToPem = (publicKey: forge.pki.Key): forge.pki.PEM => {
+    try {
+        return forge.pki.privateKeyToPem(publicKey);
+    } catch (e) {
+        throw new Error("Given publicKey is not a valid");
+    }
+};
+
+export const privateKeyFromPem = (publicKey: forge.pki.PEM): forge.pki.Certificate => {
+    try {
+        return forge.pki.privateKeyFromPem(publicKey);
+    } catch (e) {
+        throw new Error("Given publicKey is not a valid");
+    }
+};
+
 /**
  * Function that generate RSA key pair (private and public keys)
  * Function take one param that equal to rsa key size, by default is 2048.
@@ -134,4 +144,30 @@ export const publicKeyFromPem = (publicKey: forge.pki.PEM): forge.pki.Certificat
  */
 export const generateRSAKeyPair = async (rsaKeySize: number = 2048): Promise<forge.pki.KeyPair> => {
     return await forge.pki.rsa.generateKeyPair(rsaKeySize);
+};
+
+/**
+ * Function that generate ED25519 key pair (private and public keys)
+ * Function take one param that equal to seed for generating.
+ * @param {string} seed
+ * @returns {module:node-forge.pki.KeyPair}
+ */
+export const generateED25519KeyPair = (seed?: string): forge.pki.KeyPair => {
+    if (seed) {
+        seed = new forge.util.ByteBuffer(seed, "utf8");
+        return forge.pki.ed25519.generateKeyPair({ seed });
+    }
+    return forge.pki.ed25519.generateKeyPair();
+};
+
+export const checkAddress = (address: string): void => {
+    if (!address) {
+        throw new Error("Address was not provided, please set the address");
+    }
+    if (typeof address !== "string") {
+        throw new Error("Given address is not a valid");
+    }
+    if (address.search(PATTERNS.ADDRESS) === -1) {
+        throw new Error("Given address is not a valid");
+    }
 };